requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
camelot-py==0.11.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==45.0.6
cssbeautifier==1.15.4
djlint==1.36.4
dnspython==2.7.0
EditorConfig==0.17.1
email_validator==2.2.0
et_xmlfile==2.0.0

Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
ghostscript==0.8.1

google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3


google-crc32c==1.7.1

google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jinja_filters==0.1.0
jsbeautifier==1.15.4
json5==0.12.1
lxml
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.1.1
numpy
opencv-python-headless<4.10.0.84
openpyxl==3.1.5
packaging==25.0
pandas
pathspec==0.12.1
pdfminer.six
pdfplumber
pillow==11.3.0
proto-plus==1.26.1
protobuf

pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.2
pydantic_core==2.41.4
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==3.17.4
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2

regex==2025.7.34
reportlab==4.4.4
repoze.lru==0.7
requests==2.32.4
Routes==2.5.1
rsa==4.9.1
scipy
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tabulate==0.9.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
pytesseract>=0.3.10
weasyprint>=54.0.0
python-dateutil>=2.8.2
stripe
pdf2image
razorpay


